+++ org.eclipse.jdt.ls.tests/src/org/eclipse/jdt/ls/core/internal/JDTUtilsTest.java
@@ -17,11 +17,16 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.PrintStream;
 import java.net.URI;
+import java.nio.file.FileVisitOption;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Comparator;
 
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.WorkspaceJob;
 import org.eclipse.core.runtime.CoreException;
@@ -149,5 +154,21 @@ public class JDTUtilsTest extends AbstractWorkspaceTest {
 			return Throwables.getStackTraceAsString(status.getException());
 		}
 		return null;
+	}
+
+	@Test
+	public void testFakeCompilationUnit() throws Exception {
+		String tempDir = System.getProperty("java.io.tmpdir");
+		File dir = new File(tempDir, "/test/src/org/eclipse");
+		dir.mkdirs();
+		File file = new File(dir, "Test.java");
+		file.createNewFile();
+		URI uri = file.toURI();
+		JDTUtils.resolveCompilationUnit(uri);
+		IProject project = WorkspaceHelper.getProject(ProjectsManager.DEFAULT_PROJECT_NAME);
+		IFile iFile = project.getFile("/src/org/eclipse/Test.java");
+		assertTrue(iFile.getFullPath().toString() + " doesn't exist.", iFile.exists());
+		Path path = Paths.get(tempDir + "/test");
+		Files.walk(path, FileVisitOption.FOLLOW_LINKS).sorted(Comparator.reverseOrder()).map(Path::toFile).forEach(File::delete);
 	}
 }
